Added local user permissions, and logout page redirection

+++ project.py
@@ -22,7 +22,7 @@ CLIENT_ID = json.loads(
     open('client_secrets.json', 'r').read())['web']['client_id']
 
 #Connect to Database and create database session
-engine = create_engine('sqlite:///restaurantmenu.db')
+engine = create_engine('sqlite:///restaurantmenuwithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -42,6 +42,8 @@ def gconnect():
         response = make_response(json.dumps('Invalid state'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
+
+    # Obtain authorization code
     code = request.data
     try:
 
@@ -105,6 +107,15 @@ def gconnect():
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
+    # Check if user is already in DB
+    user_id = getUserID(login_session['email'])
+    
+    #If not, create the user in the DB based on the login session
+    if not user_id:
+        user_id = createUser(login_session)
+    #Finally add the user id info to the login session information
+    login_session['user_id'] = user_id
+
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
@@ -114,7 +125,37 @@ def gconnect():
     output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
     flash("you are now logged in as %s" % login_session['username'])
     print "done!"
+
     return output
+
+
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+
+@app.route('/logout')
+def showLogout():
+    if 'username' in login_session:
+        return render_template('logout.html')
+    return redirect('/')  
 
 
 @app.route('/gdisconnect')
@@ -140,6 +181,7 @@ def gdisconnect():
         del login_session['username']
         del login_session['email']
         del login_session['picture']
+        del login_session['user_id']
         response = make_response(json.dumps('Successfully disconnected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
@@ -179,7 +221,10 @@ def restaurantsJSON():
 @app.route('/restaurant/')
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+  if 'user_id' in login_session:
+      return render_template('restaurants.html', restaurants = restaurants, user=getUserInfo(login_session['user_id']))
+  else:
+      return render_template('publicrestaurants.html', restaurants = restaurants)
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
@@ -188,28 +233,32 @@ def newRestaurant():
       return redirect('/login')
 
   if request.method == 'POST':
-      newRestaurant = Restaurant(name = request.form['name'])
+      newRestaurant = Restaurant(name = request.form['name'], user_id = login_session['user_id'])
       session.add(newRestaurant)
       flash('New Restaurant %s Successfully Created' % newRestaurant.name)
       session.commit()
       return redirect(url_for('showRestaurants'))
   else:
-      return render_template('newRestaurant.html')
+      return render_template('newRestaurant.html', user=getUserInfo(login_session['user_id']))
 
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
   if 'username' not in login_session:
       return redirect('/login')
-
+  
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-      if request.form['name']:
-        editedRestaurant.name = request.form['name']
-        flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
-        return redirect(url_for('showRestaurants'))
+  if editedRestaurant.user_id != login_session['user_id']:
+      flash('You are not the creator of %s restaurant, and cannot modify it' % editedRestaurant.name)
+      return redirect(url_for('showRestaurants'))
   else:
-    return render_template('editRestaurant.html', restaurant = editedRestaurant)
+      if request.method == 'POST':
+          if request.form['name']:
+              editedRestaurant.name = request.form['name']
+              flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
+              return redirect(url_for('showRestaurants'))
+      else:
+          return render_template('editRestaurant.html', restaurant = editedRestaurant, user=getUserInfo(login_session['user_id']))
 
 
 #Delete a restaurant
@@ -219,22 +268,35 @@ def deleteRestaurant(restaurant_id):
       return redirect('/login')
 
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-    session.delete(restaurantToDelete)
-    flash('%s Successfully Deleted' % restaurantToDelete.name)
-    session.commit()
-    return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
+  if restaurantToDelete.user_id != login_session['user_id']:
+      flash('You are not the creator of %s restaurant, and cannot modify it' % restaurantToDelete.name)
+      return redirect(url_for('showRestaurants'))    
   else:
-    return render_template('deleteRestaurant.html',restaurant = restaurantToDelete)
+      if request.method == 'POST':
+          session.delete(restaurantToDelete)
+          flash('%s Successfully Deleted' % restaurantToDelete.name)
+          session.commit()
+          return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
+      else:
+          return render_template('deleteRestaurant.html',restaurant = restaurantToDelete, user=getUserInfo(login_session['user_id']))
 
 #Show a restaurant menu
 @app.route('/restaurant/<int:restaurant_id>/')
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
+#    if 'username' not in login_session:
+#        return redirect('/login')
+
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
-     
+    creator = getUserInfo(restaurant.user_id)
+    if 'user_id' in login_session: 
+        if restaurant.user_id != login_session['user_id']:
+            return render_template('publicmenu.html', items = items, restaurant = restaurant, creator = creator, user=getUserInfo(login_session['user_id']))
+        else:
+            return render_template('menu.html', items = items, restaurant = restaurant, creator = creator, user=getUserInfo(login_session['user_id']))
+    else:
+        return render_template('publicmenu.html', items = items, restaurant = restaurant, creator = creator)             
 
 
 #Create a new menu item
@@ -244,6 +306,10 @@ def newMenuItem(restaurant_id):
       return redirect('/login')
 
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  if restaurant.user_id != login_session['user_id']:
+      flash('You are not the creator of %s restaurant, and cannot modify it' % restaurant.name)
+      return redirect(url_for('showMenu', restaurant_id = restaurant.id))
   if request.method == 'POST':
       newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
       session.add(newItem)
@@ -251,16 +317,21 @@ def newMenuItem(restaurant_id):
       flash('New Menu %s Item Successfully Created' % (newItem.name))
       return redirect(url_for('showMenu', restaurant_id = restaurant_id))
   else:
-      return render_template('newmenuitem.html', restaurant_id = restaurant_id)
+      return render_template('newmenuitem.html', restaurant_id = restaurant_id, user=getUserInfo(login_session['user_id']))
 
 #Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
     if 'username' not in login_session:
-      return redirect('/login')
+        return redirect('/login')
 
     editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    
+    if restaurant.user_id != login_session['user_id']:
+        flash('You are not the creator of %s restaurant, and cannot modify it' % restaurant.name)
+        return redirect(url_for('showMenu', restaurant_id = restaurant.id))
+  
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -275,24 +346,29 @@ def editMenuItem(restaurant_id, menu_id):
         flash('Menu Item Successfully Edited')
         return redirect(url_for('showMenu', restaurant_id = restaurant_id))
     else:
-        return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
+        return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem, user=getUserInfo(login_session['user_id']))
 
 
 #Delete a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
     if 'username' not in login_session:
-      return redirect('/login')
+        return redirect('/login')
 
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
+    
+    if restaurant.user_id != login_session['user_id']:
+        flash('You are not the creator of %s restaurant, and cannot modify it' % restaurant.name)
+        return redirect(url_for('showMenu', restaurant_id = restaurant.id))
+  
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
         flash('Menu Item Successfully Deleted')
         return redirect(url_for('showMenu', restaurant_id = restaurant_id))
     else:
-        return render_template('deleteMenuItem.html', item = itemToDelete)
+        return render_template('deleteMenuItem.html', item = itemToDelete, user=getUserInfo(login_session['user_id']))
